"cart"

+++ src/app/services/cart.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
+import { BehaviorSubject, Subject } from 'rxjs';
 import {Cart} from 'src/app/model/cart.model'
 
 import {ProductbyidService} from 'src/app/services/productbyid.service'
@@ -10,22 +10,80 @@ import { Product } from '../model/product';
   providedIn: 'root'
 })
 export class CartService {
-  _obeservable: BehaviorSubject<Cart>
+  private _obeservable: BehaviorSubject<Object>
    cart={}
 
   constructor(productService: ProductbyidService) {
-    localStorage.setItem('cart',JSON.stringify(this.cart))
+    if(!this.isCartExist){
+      localStorage.setItem('cart',JSON.stringify(this.cart))
+    }
+    this. readDataFromCart()
+    this._obeservable = new BehaviorSubject(this.cart)
+   
    }
+
+   readDataFromCart(){
+     this.cart=JSON.parse(localStorage.getItem('cart'))   
+   }
+   writeDataToLocalStorage(){
+    localStorage.setItem('cart', JSON.stringify(this.cart)) 
+   }
+
+  get  getCartObservable(){
+     return this._obeservable
+   }
+
+
+   isCartExist(){
+     if(localStorage.getItem('cart')){
+       return true
+     }else{
+       return false
+     }
+   }
+   getQuantity(product: Product){
+     return this.cart[product.id] ?  +this.cart[product.id]:0
+  }
 
    
 
-   addToCart(id:string,quantity){
-     this.cart[id]=quantity;
-     localStorage.setItem("cart",JSON.stringify(this.cart))
-     console.log(this.cart);
-     
+   addToCart(product: Product){
+
+    let quantity=this.cart[ product.id];
+
+    if(quantity){
+      this.cart[product.id]=(+quantity)+1
+    }else{
+      this.cart[product.id]=1;
+    }
+
+    localStorage.setItem("cart",JSON.stringify(this.cart))
+    // this._obeservable= new BehaviorSubject(this.cart);
+    this._obeservable.next(this.cart)
+  
    
+  }
+   getFromCart(){
+     return localStorage.getItem('cart');
    }
+
+
+   addProduct(product: Product){
+
+   }
+   
+   removeProduct(product: Product){
+
+  }
+  setquantity(Product: Product,quantity: number){
+    if(quantity<1){
+      delete this.cart[Product.id];}
+      else{
+        this.cart[Product.id]=quantity
+      }
+      this.writeDataToLocalStorage();
+      this._obeservable.next(this.cart)
+    }
